+++ src/main/java/naem/server/domain/Tag.java
@@ -8,16 +8,15 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
 
 import lombok.Getter;
 import lombok.Setter;
 import naem.server.domain.member.MemberTag;
+import naem.server.domain.post.PostTag;
 
 @Entity
 @Getter
